// src/api/downloadApi/download.api.url.ts
export enum downApiUrl {
  uploadFile = '/resource/upload_file',
  resourceDownload = '/resource/download',
  resourceDownload2 = '/resource/download2',
  resourceDownload3 = '/resource/download3',
  resourceGetBatchDownloadToZip = '/resource/getBatchDownloadToZip',
  getBase64 = '/resource/get_picture_base64',

  excelMultiImportAnalysisSaveData = '/excelMultiImport/analysisSaveData',
  excelMultiImportDownFileByType = '/excelMultiImport/downFileByType',
  excelMultiImportDownloadTemplate = '/excelMultiImport/download_template', // 下载导入模板
  getViewToken = '/resource/get_view_token',
}